% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{attencoef}
\alias{attencoef}
\title{Calculate mixing length for canopy air transport}
\usage{
attencoef(hgt, PAI = 3, x = 0.5, lw = 0.05, cd = 0.2, iw = 0.5,
  phi_m = 1)
}
\arguments{
\item{hgt}{canopy height (m)}

\item{PAI}{plant area index}

\item{x}{leaf distribution angle coefficient}

\item{lw}{mean leaf width (m)}

\item{cd}{drag coefficient}

\item{iw}{relative turbulence intensity}

\item{phi_m}{diabatic correction factor}
}
\value{
dimensionless attenuation coefficient for calculating the wind height profile
in vegetation canopies.
}
\description{
Calculate mixing length for canopy air transport
}
