% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{canlw}
\alias{canlw}
\title{Calculates net outgoing longwave radiation underneath vegetated canopies}
\usage{
canlw(tc, l, ref = 0.02, skyem = 0.9)
}
\arguments{
\item{tc}{temperature (deg C)}

\item{l}{leaf area index}

\item{ref}{average leaf reflectivity in longwave spectrum (1 - thermal emissivity)}

\item{skyem}{sky emissivity}
}
\value{
Flux density of net outgoing longwave radiation emitted under leaf area \code{l} (W / m2)
}
\description{
\code{canlw} calculates the flux density (W / m^2) of net outgoing longwave
radiation underneath vegetated canopies
}
