% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{cansw}
\alias{cansw}
\title{Calculates radiation received under vegetated canopies}
\usage{
cansw(globrad, dp = NA, jd, localtime, lat, long, l, x, ref = 0.2,
  hourly = FALSE, watts = TRUE, merid = round(long/15, 0) * 15,
  dst = 0, corr = 1, tme = NA, clump = 0)
}
\arguments{
\item{globrad}{a vector of incoming shortwave radiation values (either \ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^{-2} hr^{-1}}} or \ifelse{html}{\out{W m<sup>-2</sup>}}{\eqn{W m^{-2}}})}

\item{dp}{optional numeric value or vector of values indicating the proportion of \code{globarad} that is diffuse.
If no value is provided, calculated using \code{\link[=difprop]{difprop()}}}

\item{jd}{the Julian day as returned by \code{\link[=julday]{julday()}}}

\item{localtime}{a single numeric value representing local time (decimal hour, 24 hour clock)}

\item{lat}{latitude of location (decimal degrees, -ve south of equator).}

\item{long}{longitude of location (decimal degrees, -ve west of Greenwich meridian).}

\item{l}{leaf area index (total one-sided area of leaf per unit ground area)}

\item{x}{leaf distribution angle coefficient}

\item{ref}{average reflectivity of leaves in shortwave spectrum.}

\item{hourly}{Used if dp = NA. Specifies whether values of \code{rad} are hourly (see details).}

\item{watts}{Used if dp = NA. A logical value indicating  whether the units of \code{rad} are \ifelse{html}{\out{W m<sup>-2</sup>}}{\eqn{W m^{-2}}} (TRUE) or \ifelse{html}{\out{MJ m<sup>-2</sup> hr<sup>-1</sup>}}{\eqn{MJ m^{-2} hr^{-1}}} (FALSE).}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT). Default is \code{round(long / 15, 0) * 15}}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}

\item{corr}{Used if dp = NA. An optional numeric value representing a correction to account for over- or under-estimated diffuse proportions. Values > 1 will apportion a greater ammount of total radiation as diffuse than originally calculated by the formula.}

\item{tme}{object of POSIXlt indicating times in UTC. Can be used in place of specifying
\code{jd}, \code{localtime}, \code{dst} and \code{merid}}

\item{clump}{clumpiness factor for canopy (0-1, see details)}
}
\value{
shortwave radiation received underneath vegetated canopies. Units as for globrad.
}
\description{
\code{cansw} calculates the flux density of incoming shortwave radiation underneath vegetation canopies
}
\details{
Calculated the flux density of radiation received below leaf area l, not absorbed.
I.e. ground albedo not accounted for. if \code{clump} = 0 the canopy is assumed entirely
uniform and radiation transmission is as for a turbid medium. As \code{clump} approaches 1,
the canopy is assumed to be increasingly patchy, such that a greater proportion of reaches
the ground without being obscured by leaves.
}
\examples{
l <- c(0:1000)/100
sw <- cansw(500, dp = 0.5, julday(2020, 6, 21), 12, 50, -5, l, 1)
plot(sw ~ l, type = "l")
}
\seealso{
\code{\link[=cantransdir]{cantransdir()}}, \code{\link[=cantransdif]{cantransdif()}} and \code{\link[=canlw]{canlw()}}
}
