% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{cantransdif}
\alias{cantransdif}
\title{Calculates diffuse radiation transmission through vegetated canopies}
\usage{
cantransdif(l, ref = 0.2)
}
\arguments{
\item{l}{leaf area index (total one-sided leaf area per unit ground area)}

\item{ref}{average reflectivity of leaves, usually in shortwave spectrum.}
}
\value{
the proportion of diffuse beam radiation transmitted through the canopy
}
\description{
Calculates diffuse radiation transmission through vegetated canopies
}
\seealso{
\code{\link[=cansw]{cansw()}} to calculate total shortwave radiation underneath canopies
}
