% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{cantransdir}
\alias{cantransdir}
\title{Calculates direct beam radiation transmission through vegetated canopies}
\usage{
cantransdir(l, x, sa, ref = 0.2)
}
\arguments{
\item{l}{leaf area index (total one-sided leaf area per unit ground area)}

\item{x}{leaf distribution angle coefficient}

\item{sa}{solar altitude in decimal degrees as returned by \code{\link[=solalt]{solalt()}}.}

\item{ref}{average reflectivity of leaves, usually in shortwave spectrum.}
}
\value{
the proportion of direct beam radiation transmitted through the canopy
}
\description{
Calculates direct beam radiation transmission through vegetated canopies
}
\seealso{
\code{\link[=cansw]{cansw()}} to calculate total shortwave radiation underneath canopies
}
