% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{clearskyrad}
\alias{clearskyrad}
\title{Calculates clear sky radiation}
\usage{
clearskyrad(
  tme,
  lat,
  long,
  h = 0.00697,
  tc = 15,
  pk = 101.3,
  G = 2.78,
  Ie = 1352.778,
  merid = 0,
  dst = 0
)
}
\arguments{
\item{tme}{POSIXlt object of times in UTC.}

\item{lat}{latitude of the location for which the solar azimuth is required (decimal degrees, -ve south of the equator).}

\item{long}{longitude of the location for which the solar azimuth is required (decimal degrees, -ve west of Greenwich meridian).}

\item{h}{optional specific humidities (\ifelse{html}{\out{kg kg<sup>{-1}</sup> }}{\eqn{kg kg^{-1}}}).}

\item{tc}{optional temperatures (ºC).}

\item{pk}{optional pressure (kPa).}

\item{G}{optional value describing he moisture profile in the atmosphere (per Smith 1966).}

\item{Ie}{an optional single value for extra-terrestrail radiation to permit adjustment for
sun-earth distances}

\item{merid}{optional  value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT). Default is \code{round(long / 15, 0) * 15}}

\item{dst}{optional value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}
}
\value{
expected clear-sky radioation (W/m^2.
}
\description{
Calculates clear sky radiation
}
