% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour.R
\name{converthumidity}
\alias{converthumidity}
\title{Convert between different measures of humidity}
\usage{
converthumidity(h, intype = "relative", tc = 11, pk = 101.3)
}
\arguments{
\item{h}{humidity value(s). Units as follows: specific humidity (\ifelse{html}{\out{kg kg<sup>-1</sup>}}{\eqn{kg kg^{-1}}}),
absolute humidity (\ifelse{html}{\out{kg m<sup>-3</sup> }}{\eqn{kg m^{-3}}}),
relative humidity (\\%), vapour pressure (kPa).}

\item{intype}{a character string description of the humidity type of \code{h}. One of "relative", "absolute" or "specific".}

\item{tc}{A numeric value specifying the temperature (ºC).}

\item{pk}{An optional numeric value specifying the atmospheric pressure (kPa).}
}
\value{
a list of numeric humidity values with the following components:

\code{relative} relative humidity (\\%).

\code{absolute}  absolute humidity (\ifelse{html}{\out{kg m<sup>-3</sup> }}{\eqn{kg m^{-3}}}).

\code{specific} specific humidity (\ifelse{html}{\out{kg kg<sup>-1</sup> }}{\eqn{kg kg^{-1}}}).

\code{vapour_pressure} vapour pressure (kPa).
}
\description{
converts between different humidity measures, namely
vapour pressure or relative, absolute or specific humidity.
}
\details{
This function converts between vapour pressure and specific,
relative and absolute humidity, based on sea-level pressure and
temperature. It returns a list of relative, absolute and specific
humidity and vapour pressure. If one or more of the
relative humidity values exceeds 100\\% a warning is given.
}
\examples{
converthumidity(90, 'relative', 20)
converthumidity(0.01555486, 'absolute', 20)
converthumidity(0.01292172, 'specific', 20)
}
