% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{cpair}
\alias{cpair}
\title{Calculate molar specific heat of air}
\usage{
cpair(tc)
}
\arguments{
\item{tc}{temperature (deg C)}
}
\value{
specific heat of air at constant pressure (J / mol / K)
}
\description{
Calculate molar specific heat of air
}
