% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour.R
\name{dewpoint}
\alias{dewpoint}
\title{Calculates dew point temperature}
\usage{
dewpoint(ea, tc = 11, ice = TRUE)
}
\arguments{
\item{ea}{vapour pressure (kPa)}

\item{ice}{optional logical indicating whether to return frost point temperature if output
is less than zero degrees C (TRUE = yes)}

\item{air}{temperature (degrees C)}
}
\value{
dew or frost point temperature (degrees C)
}
\description{
Calculates  dew or frost point temperature
}
\examples{
# Comparison of frost and dew point
ea <- c(10:100) / 50
tdew <- dewpoint(ea, ice = F)
tfrost <- dewpoint(ea)
plot(tfrost~ea, type = "l", col = "blue", ylim = c(-20,20))
par(new = T)
plot(tdew~ea, type = "l", col = "red", ylim = c(-20,20))
}
