% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{diabatic.approx}
\alias{diabatic.approx}
\title{Calculates approximation of diabatic correction factor above canopy}
\usage{
diabatic.approx(tc, u, H, hgt, PAI, zi)
}
\arguments{
\item{tc}{temperature (deg C)}

\item{u}{wind speed at 2 m height}

\item{H}{Heat flux (W / m^2)}

\item{hgt}{hgt of canopy (m)}

\item{zi}{height to which correction factor is wanted (m)}
}
\value{
a list with the following components:

\code{psi_m} diabatic correction factor for momentum transfer

\code{psi_h} diabatic correction factor for heat transfer
}
\description{
Calculates approximation of diabatic correction factor above canopy without
using values in current time-step only.
}
\details{
The function \code{diabatic.approx} bi-passes the need to run microclimate
models in hourly time-steps
}
\seealso{
\code{\link[=diabatic_cor]{diabatic_cor()}} \code{\link[=diabatic_cor_can]{diabatic_cor_can()}}
}
