% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{diabatic_cor}
\alias{diabatic_cor}
\title{Calculates diabatic correction factor above canopy}
\usage{
diabatic_cor(tc, pk = 101.3, H = 0, uf, zi = 2)
}
\arguments{
\item{tc}{temperature}

\item{pk}{atmospheric pressure (kPa)}

\item{H}{Heat flux (W / m^2)}

\item{uf}{friction velocity (m/s)}

\item{zi}{height to which correction factor is wanted (m)}
}
\value{
a list with the following components:

\code{psi_m} diabatic correction factor for momentum transfer

\code{psi_h} diabatic correction factor for heat transfer
}
\description{
Calculates the diabatic correction factors used in adjustment of
wind profiles and calculation of turbulent conductivity above canopy
}
