% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{diabatic_cor_can}
\alias{diabatic_cor_can}
\title{Calculates diabatic correction factor in canopy}
\usage{
diabatic_cor_can(tc, uz, z, PAI, x = 1, lw = 0.05)
}
\arguments{
\item{tc}{vector of air temperatures at canopy nodes (deg C)}

\item{uz}{vector of wind speeds at canopy nodes (m/s)}

\item{z}{vector of heights of canopy nodes (m)}

\item{PAI}{vector of plant area index values at each canopy node}

\item{x}{leaf angle coefficient}

\item{lw}{mean leaf width}
}
\value{
a list with the following components:

\code{phi_m} diabatic correction factor for momentum transfer

\code{phi_h} diabatic correction factor for heat transfer
}
\description{
Calculates the diabatic correction factors used in adjustment of
wind profiles and calculation of turbulent conductivity within the canopy
}
