% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplemodel.R
\name{.Fanim}
\alias{.Fanim}
\title{Internal function to calculate the fraction of radiation intercepted by an animal}
\usage{
.Fanim(dlen, dhgt, sa)
}
\description{
Internal function to calculate the fraction of radiation intercepted by an animal
}
