% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplemodel.R
\name{.gcanopy2}
\alias{.gcanopy2}
\title{Internal function for computing below canopy turbulent convection}
\usage{
.gcanopy2(u, zu, z1, z0, hgt, PAI)
}
\description{
Internal function for computing below canopy turbulent convection
}
