% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplemodel.R
\name{.gradabs}
\alias{.gradabs}
\title{Internal function to compute ground radiation absorption}
\usage{
.gradabs(
  Rsw,
  tc,
  skyem,
  dp,
  tme,
  pai,
  x,
  lat,
  long,
  alb,
  galb,
  slope,
  aspect,
  merid,
  dst,
  clump
)
}
\description{
Internal function to compute ground radiation absorption
}
