% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplemodel.R
\name{.gstomatal}
\alias{.gstomatal}
\title{Internal function to calculate leaf stomatal conductance}
\usage{
.gstomatal(Rsw, gsmax)
}
\description{
Internal function to calculate leaf stomatal conductance
}
