% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{.habgen}
\alias{.habgen}
\title{generates PAI and pLAI for a given habitat type}
\usage{
.habgen(
  habitat,
  lat,
  long,
  tme,
  PAIt,
  m,
  m2,
  skew,
  spread,
  pLAIo1,
  under = TRUE
)
}
\description{
generates PAI and pLAI for a given habitat type
}
