% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplemodel.R
\name{.lradabs}
\alias{.lradabs}
\title{Internal function to compute leaf radiation absorption}
\usage{
.lradabs(Rsw, tc, skyem, dp, tme, paia, x, lat, long, alb, merid, dst, clump)
}
\description{
Internal function to compute leaf radiation absorption
}
