% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplemodel.R
\name{.PenBelow}
\alias{.PenBelow}
\title{Internal function to apply Penman-Monteith equation below canopy}
\usage{
.PenBelow(
  tc,
  t0,
  tmx,
  rh,
  pk,
  gtt,
  gt0,
  gha,
  gv,
  gL,
  radabs,
  leafdens,
  Bowen,
  lims
)
}
\description{
Internal function to apply Penman-Monteith equation below canopy
}
