% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplemodel.R
\name{.PenMont}
\alias{.PenMont}
\title{Internal function to apply Penman-Monteith equation}
\usage{
.PenMont(tc, pk, ea, radabs, gHa, gs, G, Bowen, upper, swet, lims)
}
\description{
Internal function to apply Penman-Monteith equation
}
