% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{.psunlitd}
\alias{.psunlitd}
\title{Calculates the effective fraction of sunlit leaves for diffuse radiation}
\usage{
.psunlitd(l, x, clump = 0)
}
\description{
Calculates the effective fraction of sunlit leaves for diffuse radiation
}
