% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operativetemp.R
\name{Fprad}
\alias{Fprad}
\title{Calculates view factor of organism for beam radiation}
\usage{
Fprad(theta, l, d, shape = "spheroid")
}
\arguments{
\item{theta}{angle between the longitudinal axis of the animal and solar beam (degrees) as
returned by \code{\link[=thetangle]{thetangle()}}}

\item{l}{length of organism (m) along longitudinal axis}

\item{d}{diameter of organism (m)}

\item{shape}{one of \code{spheroid} (prolate spheroid), \code{cylinder1} (cylinder with flat ends) or
\code{cylinder2} (cylinder with hemisphere ends) or \code{plate} (flat plate - e.g. leaf)}
}
\value{
The view factor (ranges between zero and 1, typically ~0.25)
}
\description{
the view factor is the average flux density of beam radiation over the entire
surface of the object relative to the flux density on a flat absobing surface facing the
source
}
\details{
The view factor is used in calculation of absorbed radiation. It is effectively
the ratio of of the projected area in the direction of the solar beam to the total surface
area of the object. For a flat plate normal to the beam, the view factor is 0.5, as only have the
plate faces the beam.
}
\examples{
# caterpillar on slope (orientation unknown, so calculated by averaging across all orientations)
aspect <- c(0:359)
theta<-thetangle(alt = 50, azi = 180, slope = 20, aspect)
Fp <- Fprad(theta, 0.05, 0.01, shape = "cylinder2")
plot(Fp~aspect,type="l")
Fpm <- mean(Fp) # mean viewing angle
Fpm
}
\seealso{
\code{\link[=thetangle]{thetangle()}}, \code{\link[=Radabs]{Radabs()}}
}
