% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{gcanopy}
\alias{gcanopy}
\title{Calculates conductance under turbulent convection within the canopy}
\usage{
gcanopy(uh, z1, z0, tc1, tc0, hgt, PAI = 3, x = 0.5, lw = 0.05,
  cd = 0.2, iw = 0.5, phi_m = 1, pk = 101.3)
}
\arguments{
\item{uh}{wind speed at height of canopy top (m/s) as returned by \code{\link[microclimC:windheight_canopy]{microclimC::windheight_canopy()}}
or \code{\link[microclimC:windcanopy]{microclimC::windcanopy()}}}

\item{z1}{upper height to which conductance is wanted (m)}

\item{z0}{lower height from which conductance is wanted (m)}

\item{hgt}{height of canopy (m)}

\item{PAI}{plant area index formixing and attenuation coefficients}

\item{x}{the ratio of vertical to horizontal projections of leaf foliage}

\item{lw}{mean leaf width (m)}

\item{cd}{drag coefficient}

\item{iw}{relative turbulence intensity}

\item{phi_m}{diabatic correction factor}

\item{pk}{atmospheric pressure used for calculating moar density of air (kPA)}

\item{tc}{temperature used for calculating molar density of air (deg C)}

\item{H}{Heat flux used to determin minimum conductance, usually from previous
timestep  (W/m^2)}
}
\value{
conductance under tubulent convection within the canopy (mol/m^2/sec)
}
\description{
Calculates conductance under turbulent convection within the canopy
}
