% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{gturb}
\alias{gturb}
\title{Calculate conductance under turbulent convection above canopy}
\usage{
gturb(u, zu = 2, z1, z0 = NA, hgt, PAI = 3, tc = 15, psi_m = 0,
  psi_h = 0, zm0 = 0.004, pk = 101.3)
}
\arguments{
\item{u}{wind speed at height \code{zu} (m/s)}

\item{zu}{height of wind speed measurement (m)}

\item{z1}{upper height to which conductance is wanted (m)}

\item{z0}{lower height from which conductance is wanted (set to heat
exchange surface of canopy if NA)}

\item{hgt}{height of canopy (m)}

\item{PAI}{plant area index for determining canopy roughness length}

\item{tc}{temperature used for calculating molar density of air (deg C)}

\item{psi_m}{diabatic correction factor for momentum transfer}

\item{psi_h}{diabataic correction factor for heat transfer}

\item{zm0}{minimum surface roughness}

\item{pk}{atmospheric pressure (kPA) used for calculating molar density of air}
}
\value{
molar conductance under turbulent convection above canopy (mol/m^2/s)
}
\description{
Calculate conductance under turbulent convection above canopy
}
