% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{habitatvars}
\alias{habitatvars}
\title{Derives vegetation paramaters from habitat type}
\usage{
habitatvars(habitat, lat, long, tme, m = 20, PAIt = NA)
}
\arguments{
\item{habitat}{a character string or numeric value indicating the habitat type. See \code{\link[=habitats]{habitats()}}}

\item{lat}{a single numeric value representing the mean latitude of the location for which the solar index is required (decimal degrees, -ve south of the equator).}

\item{long}{a single numeric value representing the mean longitude of the location for which the solar index is required (decimal degrees, -ve west of Greenwich meridian).}

\item{tme}{a single value or vector of POSIXlt objects of date(s) and time(s). See details}

\item{m}{number of canopy nodes}

\item{PAIt}{optional single numeric value or vector of values of total plant area index for canopy (see details).}
}
\value{
a list with the following components:

\code{hgt} height of vegetation (m)

\code{PAI} a vector or array of Plant Area Index values (see details)

\code{x} the ratio of vertical to horizontal projections of leaf foliage

\code{lw} mean leaf width (m)

\code{cd} drag coefficient

\code{iw} a vector of length \code{m} of relative turbulence intensities

\code{hgtg} height of understory vegetation (m)

\code{zm0} roughnes length governing momentum transfer of understory vegetation (m)

\code{pLAI} a vector or array of proportions of green vegetation (see details)

\code{refls} reflectivity (shortwave radiation) of leaves

\code{refg} reflectivity (shortwave radiation) of ground

\code{refw} reflectivity (shortwave radiation) of dead vegetation and woody stems

\code{reflp} #  Reflectivity of leaves to photosynthetically active radiation

\code{vegem} thermal emissivity of vegetation

\code{gsmax} maximum stomatal conductance (mol / m^2 / s)

\code{q50} Value of PAR when photosynthetically active radiation is at 50% of
maximum (micromoles / m^2 / s^1)

\code{thickw} a vector of length \code{m} of mean thickness (m) of vegetation such that \code{thickw}
x PAI = volume of vegetation (m^3 / m^3)

\code{cpw} Specific heat capacity of woody vegetation (J / kg / K)

\code{phw} Density of woody vegetation (no air) (kg / m^3)

\code{kwood} Thermal conductivity of wood (W / m / K)

\code{clump} Clumpiness factor for canopy (0-1)
}
\description{
\code{habitatvars} generates vegetation parameters required to run the model
from habitat type.
}
\details{
if \code{tme} is a single value \code{PAI} and \code{pLAI} are vectors of length \code{m} giving
Plant Area Index values and proportions of green vegetation for each canopy node \code{m}.
If \code{tme} is a vector of times, \code{PAI} and \code{pLAI} are arrays of dimension m x number of
hours. I.e. seperate \code{PAI} and \code{pLAI} values are derived for each hour and node. If \code{tme}
is not in hourly time increments, it is conveted to hourly time increments, from 00:00 hrs
on the first day toup to and including 23:00 hrs on the final day. If \code{PAIt} is provided,
and a single value, then the length of \code{tme} is ignored and returned \code{PAI} is a vector of values for
each node. If \code{PAIt} is a vector of values of length \code{m} then seasonally adjusted values for
each node and hour of \code{tme} are returned. If \code{PAIt} is a vector of values of any length
other than \code{m} spline interpolated values for each node and hour of \code{tme} are returned.
If \code{PAIt} is NA (the default) PAI is estimated from habitat and location using \code{\link[=PAIfromhabitat]{PAIfromhabitat()}}.
}
