% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hourlyncep_convert}
\alias{hourlyncep_convert}
\title{Reformats a data frame returned by \code{\link[microclima:hourlyNCEP]{microclima::hourlyNCEP()}}}
\usage{
hourlyncep_convert(climdata, lat, long)
}
\arguments{
\item{climdata}{data.frame as returned by \code{\link[microclima:hourlyNCEP]{microclima::hourlyNCEP()}}}

\item{lat}{(declimal degrees)}

\item{long}{(decimal degrees)}
}
\value{
a data.frame of hourly weather variables (see e.g. \code{weather}).
}
\description{
Reformats a data frame returned by \code{\link[microclima:hourlyNCEP]{microclima::hourlyNCEP()}}
}
\details{
The \code{hourlyncep_convert} function reformats the data.frame returned by
\code{\link[microclima:hourlyNCEP]{microclima::hourlyNCEP()}}  for use with \code{microclimc}

The function \code{\link[microclima:hourlyNCEP]{microclima::hourlyNCEP()}} downloads the required climate and
radiation forcing data needed to run microclimate models from the NOAA-NCEP reanalysis
project and interpolates 4x daily data to hourly. Thius function reformats those data,
converting humdity and radiation values for use with the microclimc package.
}
