% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{iwgeometry}
\alias{iwgeometry}
\title{Generates relative turbulence intensity profile}
\usage{
iwgeometry(m, iwmin = 0.36, iwmax = 0.9, increasing = FALSE)
}
\arguments{
\item{m}{number of canopy nodes}

\item{iwmin}{minimum relative turbulence intensity}

\item{iwmax}{maximum relative turbulence intensity}

\item{increasing}{logical indicating whether turbulence intensity increases (TRUE) or decreases
(FALSE) with height}
}
\value{
relative turbulence intensity for each node \code{m}.
}
\description{
Generates relative turbulence intensity profile
}
\details{
default values are for a maize crop Shaw et al (1974) Agricultural Meteorology,
13: 419-425.
}
