% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{jday}
\alias{jday}
\title{Calculates the astronomical Julian day}
\usage{
jday(year, month, day, hour = 12, min = 0, sec = 0, dst = 0, tme = NA)
}
\arguments{
\item{year}{year (AD).}

\item{month}{month in numeric form (1-12).}

\item{day}{days of the month (1-31).}

\item{hour}{hours (decimal, 0-23).}

\item{min}{minutes (decimal, 0-59).}

\item{sec}{seconds (decimal, 0-59).}

\item{dst}{an optional numeric value specifying the time zone expressed as hours different from GMT (-ve to west).}

\item{tme}{POSIXlt object of times in UTC. Other inputs ignored if this is provided}
}
\value{
Julian Day. I.e. the number of days since January 1, 4713 BCE at noon UTC.
}
\description{
\code{jday} is used to calculate the astronomical Julian day (days since since January 1, 4713 BCE at noon UTC) from a given year, month and day.
}
\examples{
tme <- as.POSIXlt(0, origin = "2020-03-21 12:43", tz = "UTC")
jday(tme = tme)
jd1 <- jday(2010, 1, 31)
jd2 <- jday(2010, 1, 31, 11, 0, 0)
jd1 - jd2
}
