% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{LAIfrac}
\alias{LAIfrac}
\title{Generates fraction of green leaves profile}
\usage{
LAIfrac(m, pLAI, skew = 7)
}
\arguments{
\item{m}{number of canopy nodes}

\item{pLAI}{mean proportion of green leaf foliage}

\item{skew}{number between 0 and 10 indicating the degree of skew towards top of
canopy in green foliage (see details).}
}
\value{
vector of length \code{m} of the proportion of plant area that is green leaf area
}
\description{
Generates fraction of green leaves profile
}
\examples{
plot(LAIfrac(100,0.8, 0),type="l", ylim = c(0,1))
plot(LAIfrac(100,0.8, 5),type="l", ylim = c(0,1))
plot(LAIfrac(100,0.8, 10),type="l", ylim = c(0,1))
}
