% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour.R
\name{layercond}
\alias{layercond}
\title{Calculate canopy layer vapour conductivity}
\usage{
layercond(Rsw, gsmax, q50 = 100)
}
\description{
adjusts stomatal conductivity based on available incoming shortwave
radiation
@param Rsw incoming shortwave radiation (W / m^2)
@param gsmax maximum stomatal conductivity (mol / m^2 / sec)
@param q50 amount of photosynthetically active radiation when stomatal conductance
is at 50% of its maximum
@return stomatal conductance (mol / m^2 / sec)
@export
}
