% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{layerinterp}
\alias{layerinterp}
\title{Interpolates values from merged canopy layers}
\usage{
layerinterp(y1, y2, x1)
}
\arguments{
\item{y1}{Conductance times from ground to each merged canopy node (s)}

\item{y2}{Conductance times from ground to each unmerged canopy node (s)}

\item{x1}{value to be interpolated (temperature or vapour)}
}
\value{
interpolated value (temperature of vapour)
}
\description{
interpolates temperature of vapour concentrations from merged canopy layers
to derlive values for the original canopy layers
}
