% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{layermerge}
\alias{layermerge}
\title{Determines whether canopy air layers should be merged}
\usage{
layermerge(z, gt, hgt, timestep, ph = rep(42.24, length(z)))
}
\arguments{
\item{z}{vector of heights representing nodes at the mid-point of each canopy layer (m)}

\item{gt}{conductance by tubulent heat transfer between each node and that below it (see details) (mol / m^2 / s).}

\item{hgt}{canopy height (m)}

\item{timestep}{time step of model (s)}

\item{ph}{molar density of air as returned by \code{\link[=phair]{phair()}} (mol / m^3)}
}
\value{
a list of with the following components:

\code{mrge} a vector of integers for each layer. Those with the same number represent
layers that should be merged.

\code{zth} thickness of each (unmerged layer)

\code{u} vector of unique layer after merging
}
\description{
Determines whether the time step and heat conductance between specified air
layers within the canopy are such that temperatures would equalise and hence
layers should be merged.
}
\details{
The first value of \code{gt} is conductance between the ground and the lowest node.
Subsequent values are for between nodes, as returned by \code{\link[=gcanopy]{gcanopy()}}.
The final value is the conductance between the heighest node and a point 2 m above the
canopy representing conductance in series between the heighest node and the top of
the canopy and the top of the canopy and the air above it, the latter returned
by \code{\link[=gturb]{gturb()}}.
}
