% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{mixinglength}
\alias{mixinglength}
\title{Calculate mixing length for canopy air transport}
\usage{
mixinglength(hgt, PAI)
}
\arguments{
\item{hgt}{canopy height (m)}

\item{PAI}{plant area index}
}
\value{
mixing length (m). See details
}
\description{
Calculate mixing length for canopy air transport
}
\details{
The mixing length is us used to calculate turbulent air transport
inside vegetated canopies. It is made equivalent to the above canopy value
at the canopy surface.
}
