% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operativetemp.R
\name{OperativeT}
\alias{OperativeT}
\title{Calculates Operative Temperature}
\usage{
OperativeT(Rabs, gH, gV, tair, ea, pk = 101.3, surfwet = 1,
  em = 0.97, fluxother = 0, method = "iter")
}
\arguments{
\item{Rabs}{Absorbed long- and shortwave radiation as returned by \code{\link[=Radbs]{Radbs()}} (W / m^2)}

\item{gH}{conductance for heat (mol / m^2 / sec)}

\item{gV}{conductance for vapour (mol / m^2 / sec)}

\item{tair}{temperature of air surrounding organism (deg C)}

\item{ea}{Vapour pressur eof air surrounding organism (kPa)}

\item{pk}{Atmospheric pressure (kPa)}

\item{surfwet}{Fraction of surface area acting like a free‐water surface}

\item{em}{thermal emissivity}

\item{fluxother}{Fluxe sother than radiation, sensible and latent heat operating on
organism (W / m^2). Positive if warming organism (e.g. metabolism), negative if leaving
organism (e.g. ground heat flux)}

\item{method}{one of \code{iter}, \code{Penman} or \code{PenMon} (see details)}
}
\value{
Humid Operative Temperature of organism - i.e. steady-state temperature of
organism (deg C)
}
\description{
The function \code{OperativeT} calculates the humid operative temperature of
an organism
}
\details{
The Humid Operative Temperature is the temperature at which all fluxes operating
on the organism sum to zero. If method = \code{iter}, this temperature is calaculated iteratively
and the inputs must be single values rather than vectors. If method = \code{open}, radiation
emmited is linearized by computing radiative conductance, and the Penman linearization for
latent heat flux is used. The results obtained are similar, though not quite as accurate
as those obtained using \code{iter}.If method = \code{PenmMon}, the Penman-Monteith equation is
used, i.e. it is assumed that the temperature and humidity of the environment are influenced
by evapotransporation from the organism, such that humidity increases and the air cools.
In consequence, there is less latent heat flux, and the operative temperatures are higher.
}
\examples{
Rabs <- c(200:700)
OT1 <- 0
for (i in 1:length(Rabs)) OT1[i] <- OperativeT(Rabs[i], 0.3, 0.2, 15, 1.2, method = "iter")
OT2 <- OperativeT(Rabs, 0.3, 0.2, 15, 1.2, method = "open")
OT3 <- OperativeT(Rabs, 0.3, 0.2, 15, 1.2, method = "PenmMon")
plot(OT1 ~ Rabs, type = "l", ylim = c(0,40), lwd = 2)
par(new=T)
plot(OT2 ~ Rabs, type = "l", ylim = c(0,40), col = "blue", lwd = 2)
par(new=T)
plot(OT3 ~ Rabs, type = "l", ylim = c(0,40), col = "red", lwd = 2)
}
