% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{PAIfromhabitat}
\alias{PAIfromhabitat}
\title{Derives leaf area index, leaf geometry and canopy height from habitat}
\usage{
PAIfromhabitat(habitat, lat, long, year, meantemp = NA, cvtemp = NA,
  rainfall = NA, cvrain = NA, wetmonth = NA)
}
\arguments{
\item{habitat}{a character string or numeric value indicating the habitat type. See \code{\link[=habitats]{habitats()}}}

\item{lat}{a single numeric value representing the mean latitude of the location for which the solar index is required (decimal degrees, -ve south of the equator).}

\item{long}{a single numeric value representing the mean longitude of the location for which the solar index is required (decimal degrees, -ve west of Greenwich meridian).}

\item{year}{the year for which data are required.}

\item{meantemp}{an optional numeric value of mean annual temperature (ºC) at reference height.}

\item{cvtemp}{an optional numeric value of the coefficient of variation in temperature (K per 0.25 days) at reference height.}

\item{rainfall}{an optional numeric value mean annual rainfall (mm per year)}

\item{cvrain}{an optional numeric value of the coefficient of variation in raifall (mm per 0.25 days) at reference height.}

\item{wetmonth}{an optional numeric value indicating which month is wettest (1-12).}
}
\value{
a list with the following items:

\code{lai} hourly leaf area index values,

\code{x} the ratio of vertical to horizontal projections of leaf foliage

\code{height} the heigbht of the canopy in metres

\code{obs_time} an object of class POSIXlt of dates and times coressponding to
each value of \code{lai}
}
\description{
\code{PAIfromhabitat} generates an hourly dataset for an entire year of
leaf area index values, the ratio of vertical to horizontal projections of leaf
foliage and canopy height from habitat.
}
\details{
If no values of \code{meantemp}, \code{cvtemp}, \code{rainfall}, \code{cvrain} or \code{wetmonth} are
provided, values are obtained from \code{\link[=globalclimate]{globalclimate()}}. Variable \code{lai} is derived by fitting
a Gaussian curve parameters of which are climate and location-dependent. Functional
fits were calibrated using MODIS data (see https://modis.gsfc.nasa.gov/data/).
}
\examples{
pxh <- PAIfromhabitat("Deciduous broadleaf forest", 50, -5.2, 2015)
pxh$height
pxh$x
plot(pxh$lai ~ as.POSIXct(pxh$obs_time), type = "l", xlab = "Month",
     ylab = "LAI", ylim = c(0, 6))
}
