% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{PAIgeometry}
\alias{PAIgeometry}
\title{Generates plant area index profile}
\usage{
PAIgeometry(m, PAI, skew, spread)
}
\arguments{
\item{m}{number of canopy nodes}

\item{PAI}{total plant area index for canopy}

\item{skew}{number between 0 and 10 indicating the degree of skew towards top of
canopy in canopy foliage (see details)}

\item{spread}{positive non-zero number less than 100 indicating the degree of spread in
canopy foliage (see details)}
}
\value{
a vector of length \code{m} of plant area indices for each canopy layer
}
\description{
Generates plant area index profile
}
\details{
in specifyign \code{skew}, lower numbers indicate greater skew towards top of
canopy (5 = symetrical). In specifying \code{spread} a value of one indicates almost
all the foliage in concentrated in one canopy layer, whereas a value of 100 indicates
completely evenly spread.
}
\examples{
pai <- PAIgeometry(100, 10, 7, 70)
plot(pai, type = "l")
}
