% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimC.R
\name{paraminit}
\alias{paraminit}
\title{Initialise paramaters for first time step of model}
\usage{
paraminit(m, sm, hgt, tair, u, relhum, tsoil, Rsw)
}
\arguments{
\item{m}{number of canopy layer nodes}

\item{sm}{number of soil layers}

\item{hgt}{height of canopy (m)}

\item{tair}{air temperature at reference height (deg C)}

\item{u}{wind speed at reference height (m/s)}

\item{relhum}{relative humidity at reference height (percentage)}

\item{tsoil}{temperature of deepest soil layer. Usually ~mean annual temperature
(deg C). See details.}

\item{Rsw}{total incoming shortwave radiation (W / m^2)}
}
\value{
a list with the following elements:

\code{tc} a vector of air temperatures for each canopy layer (deg C)

\code{soiltc} a vector of airsoil temperatures for each soil layer (deg C)

\code{tleaf} a vector of leaf temperatures for each canopy layer (deg C)

\code{tabove} initially set temperature above canopy temperature, here set as \code{tair}

\code{uz} a vector of wind speeds for each canopy layer (m/s)

\code{rh} a vector of relative humidities

\code{relhum} relative humidity at 2 m above canopy (percentage)

\code{tair} air temperature at 2 m above canopy (deg C)

\code{pk} pressure at 2 m above canopu (kPA)

\code{Rabs} Absorbed radiation (W / m^2)

\code{gt} Conductivity in air of each canopy layer node (mol/m^2/sec)

\code{gv} Leaf conductivity to vapour loss for each canopy layer node  (mol/m^2/sec)

\code{gha} Conductivity between air and leaf for each canopy layer node (mol/m^2/sec)
}
\description{
Generates a set of climate and conductivity parameters for running the
first time step of the model
}
\details{
All values are approximate. Values for \code{tc} and \code{tsoil} are derived by
linear intepolation between \code{tair} and \code{tsoil}. Values for \code{Rabs} are derived from
\code{Rsw} but attenuate through the canopy. Values for \code{tleaf} are derived
from \code{tc} and \code{Rabs}. Values for \code{rh} are the same as \code{relhum}. Values for \code{gt}
\code{gv} and \code{gha} are typical for decidious woodland with wind above canopy at 2 m/s.
\code{gt} is scaled by canopy height and \code{m} (and hence distance between nodes). The first
value represents conductivity between the ground and the lowest canopy node. The last
value represents conductivity between the air at 2 m above canopy and the highest
canopy node.
}
\examples{
paraminit(20, 10, 10, 15, 80, 11, 500)

}
