% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{phair}
\alias{phair}
\title{Calculate molar density of air}
\usage{
phair(tc = 15, pk = 101.3)
}
\arguments{
\item{tc}{temperature (deg C)}

\item{pk}{atmospheric pressure (kPa)}
}
\value{
molar density of air (mol / m3)
}
\description{
Calculate molar density of air
}
