% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{psunlit}
\alias{psunlit}
\title{Calculates the fraction of sunlit leaves}
\usage{
psunlit(l, x, sa)
}
\arguments{
\item{l}{leaf area index}

\item{x}{leaf angle distribution coefficient (ratio nof horizontal to vertical projection of foliage)}

\item{sa}{solar altitude (decimal degrees)}
}
\description{
\code{psunlit} calculates the fraction of sunlit leaves for application of \code{radmult}
}
\seealso{
\code{\link[=radmult]{radmult()}}
}
