% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operativetemp.R
\name{Radabs}
\alias{Radabs}
\title{Calculates radiation absorbed by an organism}
\usage{
Radabs(
  dni,
  dif,
  si,
  tair,
  tground,
  l,
  d,
  theta,
  refsw = 0.5,
  em = 0.97,
  lwta = 0.5,
  lwtg = 1,
  galb = 0.15,
  shape = "spheroid",
  suspended = FALSE
)
}
\arguments{
\item{dni}{Flux density of beam radiation perpendicular to the solar beam (W / m2)}

\item{dif}{Flux density of beam radiation perpendicular to the solar beam (W / m2) (see \code{\link[=difprop]{difprop()}})}

\item{si}{Solar coefficient as returned by \code{\link[=solarcoef]{solarcoef()}}}

\item{tair}{temperature of air surrounding organism (deg C)}

\item{tground}{temperature of ground surface  (deg C)}

\item{l}{length or organism (m) along longitudinal axis}

\item{d}{diameter (m) of organism}

\item{theta}{angle between longitudinal axis of organism and sun as returned by \link{thetangle}}

\item{refsw}{reflectivity of organism (shortwave radiation)}

\item{em}{emissivity of organism (longwave radiation)}

\item{lwta}{fraction of longwave radiation transmitted downward (see \code{\link[=canlw]{canlw()}}.
1 - sky emissivity if no canopy cover.}

\item{lwtg}{fraction of longwave radiation transmitted upward (see \code{\link[=canlw]{canlw()}}. Determined by canopy
below organism. Ignored if \code{suspended} = FALSE.}

\item{shape}{one of \code{spheroid} (prolate spheroid), \code{cylinder1} (cylinder with flat ends) or
\code{cylinder2} (cylinder with hemisphere ends) or \code{plate} (flat plate - e.g. leaf)}

\item{suspended}{optional logical indicating whether organism is suspended in free air (single
leaves, organisms with spindly legs) or on the surface (whole canopies, organisms resting on the ground)}
}
\value{
Flux density of absorbed radiation (see details)
}
\description{
calculates the sum of long and shortwave radiation absorbed by and organism
}
\details{
The returned value is the flux density (i.e. radiation absorbed per metre squared).
To convert to total flux, multiply by surface area of organism.
}
