% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{radmult}
\alias{radmult}
\title{Calculates a multiplication factor for radiation absorbtion by sunlit leaves}
\usage{
radmult(x, sa)
}
\arguments{
\item{x}{leaf angle distribution coefficient (ratio nof horizontal to vertical projection of foliage)}

\item{sa}{solar altitude (decimal degrees)}
}
\value{
multiplication factor representing ratio of radiation intercepted by leaves relative
to a horizontal surface
}
\description{
\code{radmult} calaculates a leaf angle-dependent multiplication factor for sunlit
leaves.
}
\details{
An inclined leaf facing in the direction of the sun will tend to intercept more direct
beam radiation than a horizontal surface. This function uses an approximation of the
x-dependent leaf-angle density function as detailed in Campbell (1990) Agricultural and Forest Meteorology, 49: 173-176
to calculate multiplication factor representing ratio of radiation intercepted by leaves relative
to a horizontal surface.
}
\seealso{
\code{\link[=psunlit]{psunlit()}}
}
