% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{roughlength}
\alias{roughlength}
\title{Calculate roughness length governing momentum transfer}
\usage{
roughlength(hgt, PAI = 3, zm0 = 0.004)
}
\arguments{
\item{hgt}{canopy height (m)}

\item{PAI}{Plant area index}

\item{zm0}{minimum roughness length (relfecting ground surface roughness below canopy)}
}
\value{
momentum roughness length (m)
}
\description{
Calculate roughness length governing momentum transfer
}
