% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimC.R
\name{runcanopy}
\alias{runcanopy}
\title{Run canopy model for a single time step}
\usage{
runcanopy(climvars, previn, vegp, soilp, timestep, tme, lat, long,
  edgedist = 100, sdepth = 2, reqhgt = NA, zu = 2, theta = 0.3,
  thetap = 0.3, merid = 0, dst = 0, n = 0.6)
}
\arguments{
\item{climvars}{a of climate variables needed to run the run the model for one timestep (see dataset \code{\link[=climvars]{climvars()}})}

\item{previn}{a list of model outputs form the previous timestep as returned initially by \code{\link[=paraminit]{paraminit()}}}

\item{vegp}{a list of vegetation parameters as returned by \code{\link[=habitatvars]{habitatvars()}}}

\item{soilp}{a list of soil parameters as returned by \code{\link[=soilinit]{soilinit()}}}

\item{timestep}{length of model timestep (s)}

\item{tme}{POSIXlt object of the date and time of the current time step}

\item{lat}{latitude of location (decimal degrees)}

\item{long}{lonitude of location (decimal degrees)}

\item{edgedist}{distance to open ground (m)}

\item{sdepth}{depth of deepest soil node (m)}

\item{reqhgt}{optional height for which temperature is required (see details)}

\item{zu}{height above ground of reference climate measurements (m)}

\item{theta}{volumetric water content of upper most soil layer in current time step (m^3 / m^3)}

\item{thetap}{volumetric water content of upper most soil layer in previous time step (m^3 / m^3)}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT).}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}

\item{n}{forward / backward weighting for Thomas algorithm (see \code{\link[=Thomas]{Thomas()}})}
}
\value{
a list of of model outputs for the current timestep with the same format as \code{previn}
}
\description{
run the below-canopy model for a single timestep
}
\details{
model outputs are returned for each canopy node, with the number of canopy nodes (m)
determined by \code{previn}. Canopy nodes are spaced at equal heights throughout the canopy as in the
example. If \code{reqhgt} is set, the canopy node nearest to that height is set at the value specified.
}
\examples{
# Create initail parameters
tme <- as.POSIXlt(0, origin = "2020-05-04 12:00", tz = "GMT")
previn <- paraminit(20, 10, 10, 15, 80, 11, 500)
vegp <- habitatvars(4, 50, -5, tme, m = 20)
z<-c((1:20)-0.5)/20*vegp$hgt
soilp<- soilinit("Loam")
climavars <- list(tair=16,relhum=90,pk=101.3,u=2.1,tsoil=11,skyem=0.9,Rsw=500,dp=NA,
                 psi_h=0,psi_m=0,phi_m=0)
# Run model 100 times for current time step
for (i in 1:100) {
  plot(z ~ previn$tc, type = "l", xlab = "Temperature", ylab = "Height", main = i)
  previn <- runcanopy(climvars, previn, vegp, soilp, 60, tme, 50, -5)
}
}
