% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour.R
\name{satvap}
\alias{satvap}
\title{Calculate saturated vapour pressure}
\usage{
satvap(tc, ice = FALSE)
}
\arguments{
\item{tc}{temperature (degrees C)}

\item{ice}{optional logical indicating whether to calculate saturated vapour pressure
over ice (TRUE = Yes)}
}
\value{
saturated vapour pressure (kPa)
}
\description{
Calculates saturated vapour pressure
}
\examples{
# Plot saturated vapour pressure curve
tc <- c(-20:30)
es <- satvap(tc)
plot(es~tc, type = "l")
# Difference between saturated vapour pressure of water and ice
tc <- c(-50:0)
es1 <- satvap(tc)
es2 <- satvap(tc, ice = TRUE)
plot(es1 ~ tc, type = "l", lwd = 2, ylim = c(0,0.7), col = "red")
par(new = T)
plot(es2 ~ tc, type = "l", lwd = 2, ylim = c(0,0.7), col = "blue")
}
