% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimC.R
\name{soilinit}
\alias{soilinit}
\title{Returns soil parameters for a given soil type}
\usage{
soilinit(soiltype, timestep, m = 10, sdepth = 2, reqdepth = NA)
}
\arguments{
\item{soiltype}{one of \code{Sand}, \code{Loamy sand}, \code{Sandy loam}, \code{Loam}, \code{Silt},
\code{Silt loam}, \code{Sandy clay loam}, \code{Clay loam}, \code{Silty clay loam}, \code{Sandy clay},
\code{Silty clay} or \code{Clay}.}

\item{timestep}{length of time step (see details)}

\item{m}{number of soil nodes (see details)}

\item{sdepth}{depth of deepest soil node (m)}

\item{reqdepth}{optional depth (m) for which temperature is required (m)}
}
\value{
a list with the following items:

\code{Soil.type} description of soil type

\code{Smax} Volumetric water content at saturation (m^3 / m^3)

\code{Smin} Residual water content (m^3 / m^3)

\code{alpha} Shape parameter of the van Genuchten model (cm^-1)

\code{n} Pore size distribution parameter (dimensionless, > 1)

\code{Ksat} Saturated hydraulic conductivity (cm / day)

\code{Vq} Volumetric quartz content of soil

\code{Vm} Volumetric mineral content of soil

\code{Vo} Volumetric organic content of soil

\code{Mc} Mass fraction of clay

\code{rho} Soil bulk density (Mg / m^3)

\code{b} Shape parameter for Campbell model (dimensionless, > 1)

\code{psi_e} Matric potential (J / m^3)

\code{z} depth (m) of soil nodes (see details)
}
\description{
Returns soil parameters needed to run the microclimate model
for a given soil type
}
\details{
the depth of the soil nodes are automatically calculated so
as to ensure the top soil node is as shallow as possible, though not
shallower than 1/200 x \code{sdepth} (e.g. one cm when \code{sdepth} = two m)
and also ensuring that during a single time step, the volumetric heat capacity
of the soil would not be exceeded, so as to ensure a stable solution to soil
heat fluxes. Since volumetric heat capacity depends on water content, the
solution is derived by iteration. If \code{reqdepth} is not NA, the soil node
nearest to \code{reqdepth} is set to \code{reqdepth}.
}
\examples{
soilinit("Loam", 3600)
}
