% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{soilk}
\alias{soilk}
\title{Calculates soil heat conductivity and capacity}
\usage{
soilk(timestep, theta = 0.3, soilp)
}
\arguments{
\item{timestep}{model time step (s)}

\item{theta}{volumetric soil water fraction (m^3 / m^3)}

\item{soilp}{a list of soil parameters as returned by \code{\link[=soilinit]{soilinit()}}}
}
\value{
a list with the following components:

\code{cd} specific heat capacity of soil (J / m^3 / K)

\code{k} thermal conductance of soil (W / m^2 / K)
}
\description{
Calculates soil heat conductivity and capacity from soil properites
}
