% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour.R
\name{soilrh}
\alias{soilrh}
\title{Calculate relative humidity of soil}
\usage{
soilrh(theta, b = 5.2, Psie = -2.6, Smax = 0.419, tc = 11)
}
\arguments{
\item{theta}{volumetric soil moisture content (m^3 / m^3)}

\item{b}{soil hydraulic shape parameter}

\item{Psie}{soil water potential paramater (J / kg)}

\item{Smax}{Saturated volumetric soil moiosture content (m3 /m3)}

\item{tc}{soil temperature (deg C)}
}
\value{
soil relative humdidity expressed as fraction
}
\description{
calculates relative humidity of soil from volumetric water content

Default values are for Clay-loam
}
