% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{solalt}
\alias{solalt}
\title{Calculates the solar altitude}
\usage{
solalt(localtime, lat, long, jd, merid = round(long/15, 0) * 15, dst = 0)
}
\arguments{
\item{localtime}{local time (decimal hour, 24 hour clock).}

\item{lat}{latitude of the location for which the solar altitude is required (decimal degrees, -ve south of the equator).}

\item{long}{longitude of the location for which the solar altitude is required (decimal degrees, -ve west of Greenwich meridian).}

\item{jd}{Julian day expressed as an integer as returned by \code{\link[=jday]{jday()}}.}

\item{merid}{optional value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT). Default is \code{round(long / 15, 0) * 15}}

\item{dst}{optional value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}
}
\value{
solar altitude (decimal º).
}
\description{
\code{solalt} is used to calculate the solar altitude at any given location from the local time.
}
\examples{
# solar altitude at noon on 21 June 2010, Porthleven, Cornwall
jd <- jday (2010, 6, 21) # Julian day
solalt(12, 50.08, -5.31, jd)
}
