% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{solarcoef}
\alias{solarcoef}
\title{Calculates the solar coefficient}
\usage{
solarcoef(slope, aspect, localtime, lat, long, jd, merid = round(long/15,
  0) * 15, dst = 0, horizon = TRUE)
}
\arguments{
\item{slope}{slopes angle (decimal º)}

\item{aspect}{aspect angle (decimal º)}

\item{localtime}{local time (decimal hour, 24 hour clock).}

\item{lat}{latitude of the location for which the coefficientis required (decimal degrees, -ve south of the equator).}

\item{long}{longitude of the location for which the solar index is required (decimal degrees, -ve west of Greenwich meridian).}

\item{jd}{integer representing the Julian day as returned by \code{\link[=julday]{julday()}}.}

\item{merid}{optional value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT). Default is \code{round(long / 15, 0) * 15}}

\item{dst}{optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}

\item{horizon}{optinal logical indicating whether to set coefficient to zero if solar altitude < 0.}
}
\value{
the proportion of direct beam radiation incident on an inclined surface.
}
\description{
\code{solarcoef} is used to calculate the proportion of direct beam radiation
incident on an inclined surface at a specified time and location.
}
\examples{
library(raster)
jd <- julday (2010, 6, 21) # Julian day
lt <- c(0:2400) / 100
si <- solarcoef(0, 0, lt, 50, -5, jd)
plot(si ~ lt, type = "l")
}
\seealso{
the microclimafunction \code{\link[microclima:solarindex]{microclima::solarindex()}} can be used to derive the solar
coeffficient for cells of a digital elevation dataset accounting for topographic shading.
}
