% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{spinup}
\alias{spinup}
\title{Model spin-up for first time-step}
\usage{
spinup(climdata, soiltype, habitat, lat, long, m, sm = 10,
  edgedist = 100, reqhgt = NA, sdepth = 2, zu = 2, theta = 0.3,
  thetap = 0.3, merid = 0, dst = 0, n = 0.6, plotout = TRUE,
  steps = 200)
}
\arguments{
\item{climdata}{a data.frame of climate variables (see e.g. \code{weather})}

\item{soiltype}{one of \code{Sand}, \code{Loamy sand}, \code{Sandy loam}, \code{Loam}, \code{Silt},
\code{Silt loam}, \code{Sandy clay loam}, \code{Clay loam}, \code{Silty clay loam}, \code{Sandy clay},
\code{Silty clay} or \code{Clay}.}

\item{habitat}{a integer or character string specifying the habitat type (see \code{habitats})}

\item{lat}{Latitude (decimal degrees)}

\item{long}{Longitude (decimal degrees, negative west of Greenwich meridion)}

\item{m}{number of canopy nodes}

\item{sm}{number of soil nodes}

\item{edgedist}{distance to open ground (m)}

\item{reqhgt}{optional height for which temperature is required (see details)}

\item{sdepth}{depth of deepest soil node (m)}

\item{zu}{height above ground of reference climate measurements (m)}

\item{theta}{volumetric water content of upper most soil layer in current time step (m^3 / m^3)}

\item{thetap}{volumetric water content of upper most soil layer in previous time step (m^3 / m^3)}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT).}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}

\item{n}{forward / backward weighting for Thomas algorithm (see \code{\link[=Thomas]{Thomas()}})}

\item{plotout}{optional logical indicating whether to a plot a profile of temperatures
upon completion.}
}
\value{
a list of model outputs as for \code{\link[=paraminit]{paraminit()}} or \code{\link[=runonestep]{runonestep()}}
}
\description{
\code{spinup} runs the model repeatedly using data form the first time-step
for a set number of steps to ensure initial conditions are stable and
appropriate soil temperatures are set.
}
\details{
If \code{reqhgt} is set, and below the height of the canopy, the canopy node nearest
to that height is set at the value specified. The returned value \code{tabove} is then the
temperature at the top of the canopy. If \code{reqhgt} is above canopy, nodes are calculated
automatically, but \code{tabove} is the temperature at height \code{reqhgt}. If \code{reqhgt} is
negative, the soil node nearest to that height is set at the value specified.
}
