% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplemodel.R
\name{Tcanopy}
\alias{Tcanopy}
\title{Compute below-canopy air temperature}
\usage{
Tcanopy(
  climdata,
  z,
  lat,
  long,
  hgt,
  pai,
  paia = NA,
  x = 1,
  gsmax = 0.33,
  lw = 0.05,
  slope = 0,
  aspect = 0,
  alb = 0.23,
  galb = 0.15,
  zu = 2,
  Bowen = 1.5,
  upper = 25,
  umin = 0.5,
  merid = 0,
  dst = 0,
  clump = 0,
  method = "C",
  lims = TRUE
)
}
\arguments{
\item{climdata}{a dataframe of hourly weather data formated and with units
as per the internal dataset \code{climdata}}

\item{z}{the height (below canopy) for which temperature estimates are required (m)}

\item{lat}{the latitude of the location (decimal degrees)}

\item{long}{the longitude of the location (decimal degrees)}

\item{hgt}{canopy height (m)}

\item{pai}{the total one sided area of canopy elements per unit ground area (see details)}

\item{paia}{optionally, the total one sided area of canopy elements per unit ground area above z (see details)}

\item{x}{leaf distribution angle coefficient}

\item{gsmax}{maximum stomatal conductance of leaves (mol / m^2 / s)}

\item{lw}{average leaf width (m)}

\item{slope}{the slope of the underlying ground surface (decimal degrees). Ignored if \code{method} = S.}

\item{aspect}{the aspect of the underlying ground surface (decimal degrees). Ignored if \code{method} = S.}

\item{alb}{the albedo of the canopy surface (either the combined ground and
canopy albedo if \code{method = S} or just the canopy albedo)}

\item{galb}{ground surface albedo}

\item{zu}{the height above ground of wind speeds in \code{climdata} (m)}

\item{Bowen}{Optional parameter specifying the Bowen Ratio of the surface. Used to improve estimates
of temperature in application of the Penman-Monteith equation}

\item{upper}{optional upper limit to temperature offset (difference between reference
and canopy surface temperature cannot exceed this value). Ignored if \code{lims} = FALSE}

\item{umin}{optional minimum wind speed for computing conductances (avoids conductances being too low)}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT)}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if merid = 0).}

\item{clump}{clumpiness factor (0-1, see details)}

\item{method}{if \code{S} treats the vegetation surface as a flat surface (see details)}

\item{lims}{optional logical indicating whether to limit temperatures by \code{upper} and dewpoint temperature}
}
\value{
a list of the following elements:

\code{tleaf} a vector of leaf surface temperatures (deg C)

\code{tn} a vector of air temperatures (deg C)

\code{rh} a vector of relative humidities (percentage)
}
\description{
Function to compute below canopy air temperatures
}
\details{
if \code{pai} is unknown it can be estimated as -ln(1-fractional
canopy cover). If \code{paia} is unspecified vertically uniform foliage distribution is assumed.
if \code{clump} = 0 the canopy is assumed entirely uniform
and radiation transmission is as for a turbid medium. As \code{clump}
approaches 1, the canopy is assumed to be increasingly patchy, such
that a greater proportion of reaches the ground without being obscured
by leaves. If \code{method} = S, the radiation intercepted by the canopy is assumed to
be that for a flat surface. If \code{method} is not S, the radiation absorption
by the ground and canopy surface are computed separately accounting for the
inclination of the ground surface and the distribution of leaf angles.
}
\examples{
# Compute temperatures below canopy
Tln<-Tcanopy(climdata,0.25,50.2178,-5.32656,hgt=0.5,pai=2,x=1,gsmax=0.33,alb=0.23)
plot(Tln$tn,type="l") # air temperature
plot(Tln$tleaf,type="l") # leaf temperature
}
