% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplemodel.R
\name{terraincoef}
\alias{terraincoef}
\title{Compute terrain influencing factor}
\usage{
terraincoef(climatedata, lat, long, pai, slope, aspect, merid = 0, dst = 0)
}
\arguments{
\item{lat}{the latitude of the location (decimal degrees)}

\item{long}{the longitude of the location (decimal degrees)}

\item{pai}{the total one sided area of canopy elements per unit ground area (see details)}

\item{slope}{the slope of the underlying ground surface (decimal degrees). Ignored if \code{method} = S.}

\item{aspect}{the aspect of the underlying ground surface (decimal degrees). Ignored if \code{method} = S.}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT)}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if merid = 0).}

\item{climdata}{a dataframe of hourly weather data formated and with units
as per the internal dataset \code{climdata}}
}
\value{
a single numeric value giving a terrain influencing factor
}
\description{
Function to compute terrain influencing factor
}
\examples{
# Compute influencing factor of an inclined surface
iH<-terraincoef(climatedata,50.2178,-5.32656,1,30,180)
# Compute influencing factor of a flat surface
iRef<-terraincoef(climatedata,50.2178,-5.32656,1,0,0)
# Compute ratio
iR<-iH/iRef
iR
}
