% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operativetemp.R
\name{thetangle}
\alias{thetangle}
\title{Calculates angle between organism and sun}
\usage{
thetangle(alt, azi, slope = 0, aspect = 180, plate = FALSE)
}
\arguments{
\item{alt}{solar altitude (degrees) as returned by \code{\link[=solalt]{solalt()}}}

\item{azi}{solar azimuth (degrees) as returned by \link{solazi}}

\item{slope}{slope angle (degrees) of organism's longitudinal axis (see details)}

\item{aspect}{aspect (degrees relative to north) of an organism's longitudinal axis (see details)}

\item{shape}{logical indicating whether organism is flat (leaf) or not (see details)}
}
\value{
angle between organism and sun (degrees)
}
\description{
Calculates angle between organism's longitudinal axis and the direction
of the solar beam
}
\details{
Slope and aspect values are relative to the horizontal and north respectively. E.g. an upright
human has a slope of 90 degrees, whereas a caterpillar facing down-slope as the same slope
and apsect as the underlying terrain. Shape must be specified, as for a flat plate, \code{azi}
has no effect on the angle between the organism and sun when \code{slope} is 0, but for spheres
and cylinders \code{azi} has no effect on the angle between organism and sun when \code{slope} is 90.
90 degrees for a human, 0 for an organism on a flat surface. The function \code{thetangle} is used
in the calculation of view factors for direct beam radiation and absorbed radiation and if the orientation of the
organism is unknown, absorbed radiation can be calculated by assuming  no direction bias
in orientation and deriving results numerically as in the example for \code{\link[=Fprad]{Fprad()}}.
}
\examples{
alt <- c(0:90) # solar altitude
# Angle for upright human
theta <- thetangle(alt, azi = 0, slope = 90, aspect = 0)
plot(theta ~ alt, type = "l")
}
\seealso{
\code{\link[=Fprad]{Fprad()}}, \code{\link[=Radabs]{Radabs()}}
}
