% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{thickgeometry}
\alias{thickgeometry}
\title{Generates vegetation thickness profile}
\usage{
thickgeometry(m, mthick, thmax, thmin)
}
\arguments{
\item{m}{number of canopy nodes}

\item{mthick}{mean thickness of vegetation}

\item{thmax}{mean thickness of lowest portion of canopy}

\item{thmin}{mean thickness of upper portion of canopy}
}
\value{
a vector of thicknesses for each canopy node (m)
}
\description{
\code{thickgeometry} is used is to calculate the mean thickness of leaves
and woody vegetation such that volume = Plant Area Index * thickness
}
\examples{
pai <- PAIgeometry(1000, 10, 7, 70)
thick <- thickgeometry(1000,0.4,0.7,0.1)
hgt <- 25
dens <- thick * pai * length(pai) / hgt
z <- c(1:1000) / 40
par(mar=c(5, 5, 2, 2))
plot(z ~ dens, type = "l",  cex.axis = 1.5, cex.lab = 1.5,
    xlab = expression(paste("Volume ",(~m^3 / m^3))),
    ylab = "Height (m)")
}
