% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimC.R
\name{Thomas}
\alias{Thomas}
\title{Thomas algorithm for solving simultanious heat fluxes}
\usage{
Thomas(tc, tmsoil, tair, k, cd, f = 0.6, X = 0)
}
\arguments{
\item{tc}{vector of soil and air temperatures (deg C) from previous timestep (see details)}

\item{tmsoil}{temperature (deg C) of deepest soil layer. Typically mean annual temperature}

\item{tair}{air temperature at reference height 2 m above canopy in current time step (deg C)}

\item{k}{vector of thermal conductances between layers (W / m^2 / K) (see details)}

\item{cd}{thermal heat capacity of layers (W / m^2 / K)}

\item{f}{forward / backward weighting of algorithm (see details)}

\item{X}{vector of temperatures to be added resulting from e.g. leaf heat fluxes or radiation
absorbed by top soil layer}
}
\value{
a vector of temperatures (deg C) for each soil / air layer for current time step.
}
\description{
\code{Thomas} implements the Thomas algorithm for solving simultanious heat
fluxes between soil / air layers.
}
\details{
\code{tc} must be ordered with reference air temperature first and the soil tmeperature
of the  deepest layer last. The  weighting factor \code{f}  may range from 0 to 1. If \code{f} = 0,
the flux is determined by the temperature difference at the beginning of the time step.
If \code{f} = 0.5, the average of the old and new temperatures is used to compute heat flux.
If \code{f} = 1, fluxes are computed using only the new temperatures. The best value to use
for \code{f} is determined by considerations of numerical stability and accuracy and experimentation
may be required. If \code{f} = 0  more heat transfer between nodes is predicted than would actually
occur, and can therefore become unstable if time steps are too large. When \code{f} > 0.5,
stable solutions will always be obtained, but heat flux will be underestimated. The
best accuracy is obtained with \code{f} around 0.4, while best stability is at \code{f} = 1.
A typical compromise is \code{f} = 0.6.
}
