% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{windsim}
\alias{windsim}
\title{Simulates wind gusts}
\usage{
windsim(
  u,
  timestepin = 3600,
  timestepout = 1,
  gustduration = 30,
  stability = 2
)
}
\arguments{
\item{u}{a vector of wind speeds (m/s)}

\item{timestepin}{duration of time interval between sucessive values of u (s)}

\item{timestepout}{duration of time interval between sucessive values of simulated wind speed (s)}

\item{gustduration}{mean duration of wind gusts}

\item{stability}{a stability value indicating how variable wind speeds are (1 = very gusty, 3 = fairly stable)}
}
\value{
a vector os simulated wind speeds, incoporating gustiness.
}
\description{
\code{winsim} simulates a hih temporal resoltuion time series of wind speeds
that it includes wind gust
}
\examples{
u <- c(3.6, 6.2, 8.2, 9.3)
ws <- windsim(u)
par(mfrow = c(2, 1))
plot(u, type = "l", ylim = c(0,30))
plot(ws, type = "l", ylim = c(0,30))

}
