% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{zeroplanedis}
\alias{zeroplanedis}
\title{Calculate zero plane displacement}
\usage{
zeroplanedis(hgt, PAI = 3)
}
\arguments{
\item{hgt}{canopy height (m)}

\item{PAI}{Plant area index}
}
\value{
zero plane displacement (m)
}
\description{
Calculate zero plane displacement
}
