% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{gcanopy}
\alias{gcanopy}
\title{Calculates conductance under turbulent convection within the canopy}
\usage{
gcanopy(
  uh,
  z1,
  z0,
  tc1,
  tc0,
  hgt,
  PAI = 3,
  x = 0.5,
  lw = 0.05,
  cd = 0.2,
  iw = 0.5,
  phi_m = 1,
  pk = 101.3
)
}
\arguments{
\item{uh}{wind speed at height of canopy top (m/s) as returned by \code{\link[microclimc:windprofile]{microclimc::windprofile()}}}

\item{z1}{upper height to which conductance is wanted (m)}

\item{z0}{lower height from which conductance is wanted (m)}

\item{tc1}{temperature of upper layer (dec C) usually in previous timestep}

\item{tc0}{temperature of lower layer (dec C) usually in previous timestep}

\item{hgt}{height of canopy (m)}

\item{PAI}{plant area index for mixing and attenuation coefficients}

\item{x}{the ratio of vertical to horizontal projections of leaf foliage}

\item{lw}{mean leaf width (m)}

\item{cd}{drag coefficient}

\item{iw}{relative turbulence intensity}

\item{phi_m}{diabatic correction factor}

\item{pk}{atmospheric pressure used for calculating moar density of air (kPA)}
}
\value{
conductance under tubulent convection within the canopy (mol/m^2/sec)
}
\description{
Calculates conductance under turbulent convection within the canopy
}
