% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{gforcedfree}
\alias{gforcedfree}
\title{Calculates forced or free laminer conductance}
\usage{
gforcedfree(d, u, tc, dtc, pk = 101.3)
}
\arguments{
\item{d}{chacteristic dimension of surface (m)}

\item{u}{wind speed (m/s)}

\item{tc}{temperature (deg C)}

\item{dtc}{estimate of temperature differences of surface and air, e.g.
from previous time step (see details)}

\item{pk}{atompsheric pressure (KPa). used for calculating molar density
of air.}
}
\value{
conductance (mol / m^2 / sec)
}
\description{
Calculates forced or free laminer conductance
}
\details{
Calculates conductance under forced and free convection and selects
whichever is greater. For conductance under free convection an estimate of
the temperature difference between the air and  surface is needed (usually
from the previous timestep)
}
\examples{
# As function of `d`
d <- c(1:1000) / 100
g <- gforcedfree(d, 2, 11, 2)
plot(g~d, type = "l")
# As function of `u`
u <- c(1:200)/100
g<-gforcedfree(0.1,u,11,4)
plot(g~u, type = "l")
}
