% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{habitatvars}
\alias{habitatvars}
\title{Derives vegetation paramaters from habitat type}
\usage{
habitatvars(habitat, lat, long, tme, m = 20)
}
\arguments{
\item{habitat}{a character string or numeric value indicating the habitat type. See \code{\link[=habitats]{habitats()}}}

\item{lat}{a single numeric value representing the mean latitude of the location for which the solar index is required (decimal degrees, -ve south of the equator).}

\item{long}{a single numeric value representing the mean longitude of the location for which the solar index is required (decimal degrees, -ve west of Greenwich meridian).}

\item{tme}{POSIXlt object of date and time}

\item{m}{number of canopy nodes}
}
\value{
a list with the following components:

\code{hgt} height of vegetation (m)

\code{PAI} a vector of length \code{m} of Plant Area Index values

\code{x} the ratio of vertical to horizontal projections of leaf foliage

\code{lw} mean leaf width (m)

\code{cd} drag coefficient

\code{iw} a vector of length \code{m} of relative turbulence intensities

\code{hgtg} height of understory vegetation (m)

\code{zm0} roughnes length governing momentum transfer of understory vegetation (m)

\code{pLAI} a vector of length \code{m}  of the proportion of green vegetation

\code{refls} reflectivity (shortwave radiation) of leaves

\code{refg} reflectivity (shortwave radiation) of ground

\code{refw} reflectivity (shortwave radiation) of dead vegetation and woody stems

\code{reflp} #  Reflectivity of leaves to photosynthetically active radiation

\code{vegem} thermal emissivity of vegetation

\code{gsmax} maximum stomatal conductance (mol / m^2 / s)

\code{q50} Value of PAR when photosynthetically active radiation is at 50% of
maximum (micromoles / m^2 / s^1)

\code{thickw} a vector of length \code{m} of mean thickness (m) of vegetation such that \code{thickw}
x PAI = volume of vegetation (m^3 / m^3)

\code{cpw} Specific heat capacity of woody vegetation (J / kg / K)

\code{phw} Density of woody vegetation (no air) (kg / m^3)

\code{kwood} # thermal conductivity of wood (W / m / K)
}
\description{
\code{habitatvars} generates vegetation parameters required to run the model
from habitat type.
}
