% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{mixinglength}
\alias{mixinglength}
\title{Calculate mixing length for canopy air transport}
\usage{
mixinglength(hgt, PAI, x, lw)
}
\arguments{
\item{hgt}{canopy height (m)}

\item{PAI}{plant area index}

\item{x}{the ratio of vertical to horizontal projections of leaf foliage}

\item{lw}{mean leaf width (m)}
}
\value{
mixing length (m). See details
}
\description{
Calculate mixing length for canopy air transport
}
\details{
The mixing length is us used to calculate tubulent air transport
inside vegetated canopies. It is the mean free pathway, equivelent to the
mean leaf-air distance. The leaf distribution angle is used to asses whether
leaves are more like grasses or squares (assumed to be like squares if less than
1 and grasses if more than one).
}
