% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plotresults}
\alias{plotresults}
\title{Function to plot temperature profile}
\usage{
plotresults(modelout, vegp, climvars, i = "")
}
\arguments{
\item{modelout}{a list of model outputs as returned by [microclimc::runonestep() or \code{\link[microclimc:runmodel]{microclimc::runmodel()}}]}

\item{vegp}{a list of vegetation parameters as returned by \code{\link[=habitatvars]{habitatvars()}}}

\item{climvars}{a list of climate variables. See example for \code{\link[microclimc:runonestep]{microclimc::runonestep()}}}

\item{i}{optional title for plot. Usually the model iteration.}
}
\description{
\code{plotresults} optionally plots the temperature profile after running
\code{\link[microclimc:spinup]{microclimc::spinup()}} or while running the model to keep track of progress. It can also be used
to plot the results after running the model
}
