% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{roughlength}
\alias{roughlength}
\title{Calculate roughness length governing momentum transfer}
\usage{
roughlength(
  hgt,
  PAI,
  zm0 = 0.003,
  cdl = 7.5,
  CR = 0.3,
  psih = 0.193,
  umx = 0.3
)
}
\arguments{
\item{hgt}{canopy height (m)}

\item{PAI}{Plant area index}

\item{zm0}{substrate-surface drag coefficient}

\item{cdl}{Control parameter scaling d/h (from Raupach 1994)}

\item{CR}{roughness-element drag coefficient}

\item{psih}{parameter characterizes the roughness sublayer depth}

\item{umx}{Maximum ratio of wind velocity to friction velocity}
}
\value{
momentum roughness length (m)
}
\description{
Calculate roughness length governing momentum transfer
}
\references{
Raupach, M.R. (1994) Simplified expressions for vegetation roughness length and zero-plane
displacement as functions of canopy height and area index. Boundary-Layer Meteorology 71: 211-216.
}
