% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimC.R
\name{runmodel}
\alias{runmodel}
\title{Runs the microclimate model over time}
\usage{
runmodel(climdata, vegp, soilp, lat, long, edgedist = 100, reqhgt = NA,
  sdepth = 2, zu = 2, theta = 0.3, thetap = 0.3, merid = 0,
  dst = 0, n = 0.6, steps = 200, plotout = TRUE, plotsteps = 100,
  tsoil = NA)
}
\arguments{
\item{climdata}{a data.frame of climate variables (see e.g. \code{weather})}

\item{vegp}{a list of vegetation parameters as returned by \code{\link[=habitatvars]{habitatvars()}}}

\item{soilp}{a list of soil parameters as returned by \code{\link[=soilinit]{soilinit()}}}

\item{lat}{Latitude (decimal degrees)}

\item{long}{Longitude (decimal degrees, negative west of Greenwich meridion)}

\item{edgedist}{distance to open ground (m)}

\item{reqhgt}{optional height for which temperature is required (see details)}

\item{sdepth}{depth of deepest soil node (m)}

\item{zu}{height above ground of reference climate measurements (m)}

\item{theta}{volumetric water content of upper most soil layer in current time step (m^3 / m^3)}

\item{thetap}{volumetric water content of upper most soil layer in previous time step (m^3 / m^3)}

\item{merid}{an optional numeric value representing the longitude (decimal degrees) of the local time zone meridian (0 for GMT).}

\item{dst}{an optional numeric value representing the time difference from the timezone meridian (hours, e.g. +1 for BST if \code{merid} = 0).}

\item{n}{forward / backward weighting for Thomas algorithm (see \code{\link[=Thomas]{Thomas()}})}

\item{steps}{number of iterations over which to run \code{\link[=spinup]{spinup()}}}

\item{plotout}{optional logical indicating whether to a plot a profile of temperatures
upon completion.}

\item{plotsteps}{number of iterations run before resuls plotted if \code{plotout} set to TRUE}

\item{tsoil}{optional stable temperature of the deepest soil layer (see details)}
}
\value{
a data.frame with the following elements:

\code{obs_time} POSIXlt object of times associated wiht eahc output

\code{reftemp} air temperature at reference height - i.e. \code{climdata$temp}

\code{tout} air or soil temperature (see details)

\code{tleaf} Leaf temperature (see details)
}
\description{
\code{runmodel} is used to run the full model over time
}
\details{
If \code{reqhgt} is set, and below the height of the canopy, the canopy node nearest
to that height is set at the value specified. The returned values \code{tout} and \code{tleaf} are
then the air and leaf temperatures at that height. If \code{reqhgt} is above canopy, nodes are
calculated automatically, and \code{tout} is the temperature at height \code{reqhgt} and \code{tleaf} is
the mean leaf temperature. If \code{reqhgt} is negative, the soil node nearest to that height
is set at the value specified, \code{tout} is soil temperature at that node and \code{tleaf} is
the mean leaf temperature. The parameter \code{tsoil} is the temperature of soil below \code{sdepth},
which is assumed constant. If \code{tsoil} is not provided, it is assigned a value equivelent
to mean of \code{climdata$temp}.
}
\examples{
tme<-as.POSIXlt(weather$obs_time, format = "\%Y-\%m-\%d \%H:\%M", tz = "UTC")
vegp <- habitatvars(4, lat = 50, long = -5, tme, m = 20)
soilp<- soilinit("Loam")
dataout <- runmodel(weather, vegp, soilp, lat = 50, long = -5)
plot(dataout$tleaf, type = "l")
}
