% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{soilk}
\alias{soilk}
\title{Calculates soil heat conductivity and capacity}
\usage{
soilk(timestep, m, theta = 0.3, frm = 0.3, frq = 0.3, frc = 0.01,
  rho = 2.65)
}
\arguments{
\item{timestep}{model time step (s)}

\item{m}{number of soil layers}

\item{theta}{volumetric soil water fraction (m^3 / m^3)}

\item{frm}{volumetric soil mineral fraction (m^3 / m^3)}

\item{frq}{volumetric soil quartz fraction (m^3 / m^3)}

\item{frc}{mass fraction of clay (kg / kg)}

\item{rho}{bulk density of soil (Mg / m^3)}
}
\value{
a list with the following components:

\code{cd} specific heat capacity of soil (J / m^3 / K)

\code{k} thermal conductance of soil (W / m^2 / K)
}
\description{
Calculates soil heat conductivity and capacity from soil properites
}
